feat: pca switching scaffold
